Split controllers/remote_build/submit_job.py into a package
