+++ esphome_device_builder/controllers/remote_build/submit_job/__init__.py
@@ -1,0 +1,5 @@
+"""Receiver-side ``submit_job`` flow for the remote-build peer-link."""
+
+from .controller import SubmitJobReceiver
+
+__all__ = ["SubmitJobReceiver"]
