+++ esphome_device_builder/controllers/remote_build/submit_job/_extract_window.py
@@ -7,7 +7,7 @@ from collections.abc import Coroutine
 from functools import partial
 from typing import Any
 
-from ...helpers.async_ import drain_tasks, log_task_exit
+from ....helpers.async_ import drain_tasks, log_task_exit
 
 _Key = tuple[str, str]
